+++ pages/api/blogs/comments/[commentId].js
@@ -19,7 +19,7 @@ export default async function handler(req, res) {
             where: { id: parseInt(commentId) },
         });
     
-        if (!comment) {
+        if (!comment || comment.deleted || comment.hidden) {
             return res.status(404).json({ error: 'Comment not found' });
         }
 
@@ -102,7 +102,6 @@ export default async function handler(req, res) {
 
         res.status(200).json(updatedComment);
     } catch (error) {
-        console.error("Error processing comment vote:", error);
         return res.status(500).json({ error: 'Could not update comment ratings' });
     }
 }
